apps/backend/requirements.txt
# Core FastAPI & Database
fastapi==0.104.1
uvicorn==0.24.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# Database Migrations
alembic==1.13.1

# Authentication & Security
python-jose==3.3.0
python-multipart==0.0.6
passlib==1.7.4
bcrypt==4.3.0

# Validation & Serialization
pydantic==2.11.7
pydantic_core==2.33.2
email-validator==2.1.0

# API Enhancement
slowapi==0.1.9
websockets==15.0.1

# File Handling & Storage
aiofiles==24.1.0
pillow==10.4.0

# Background Tasks & Queue
celery==5.3.7
redis==5.2.0

# HTTP Client & Utils
httpx==0.25.2
httptools==0.6.4

# Data Processing & Analysis
pandas==2.2.2
numpy==2.1.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
factory-boy==3.3.0
Faker==37.4.0

# Development Tools
black==23.12.1
flake8==7.0.0
coverage==7.9.2

# System Dependencies (keep existing)
annotated-types==0.7.0
anyio==3.7.1
certifi==2025.7.9
cffi==1.17.1
click==8.2.1
cryptography==45.0.5
Deprecated==1.2.18
dnspython==2.7.0
ecdsa==0.19.1
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
idna==3.10
iniconfig==2.1.0
limits==5.4.0
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pluggy==1.6.0
pyasn1==0.6.1
pycodestyle==2.11.1
pycparser==2.22
pyflakes==3.2.0
PyYAML==6.0.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uvloop==0.21.0
watchfiles==1.1.0
wrapt==1.17.2
